+++ desktop/apps/auction2/client.js
@@ -6,7 +6,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import auctionReducer from 'desktop/apps/auction2/reducers'
 import configureStore from 'desktop/apps/auction2/utils/configureStore'
-import { rehydrateClient } from 'desktop/components/react/utils/render_react_layout'
+import { rehydrateClient } from 'desktop/components/react/utils/renderReactLayout'
 
 // Rehydrate data from Server
 const bootstrapData = rehydrateClient(window.__BOOTSTRAP__)
