+++ desktop/apps/auctions2/routes.js
@@ -1,6 +1,6 @@
 import AuctionReminders from 'desktop/components/auction_reminders/fetch.coffee'
 import Index from './templates/index.jsx'
-import renderReactLayout from 'desktop/components/react/utils/render_react_layout'
+import renderReactLayout from 'desktop/components/react/utils/renderReactLayout'
 
 export function index (req, res, next) {
   const layout = renderReactLayout({
