+++ desktop/components/react/utils/renderTemplate.js
